refact: #55 move store from src/utils/store to src/modules
