+++ frontend/package.json
@@ -96,10 +96,10 @@
       "!src/constants/*",
       "!src/index.jsx",
       "!src/serviceWorker.js",
-      "!src/utils/store/index.js",
+      "!src/modules/index.js",
+      "!src/utils/api/index.js",
       "!src/utils/dateTimeHelpers.js",
-      "!src/styles.js",
-      "!src/utils/api/index.js"
+      "!src/styles.js"
     ]
   },
   "proxy": "http://127.0.0.1:8000/"
