+++ frontend/src/modules/notification/index.test.js
@@ -1,5 +1,5 @@
 import axios from '../utils/api';
-import store from '../utils/store';
+import { store } from '../modules';
 import { mockNotifications } from '../constants';
 import notiReducer, * as actionCreators from './notification';
